refactor:#20 모든 병원 위치 가져오는 로직을 훅으로 분리
